transcoding_function/requirements.txt
google-cloud-video-transcoder==1.4.0
